Use Django database router API (#162)
-  `reactpy_django.database.Router` must now be registered in `settings.py:DATABASE_ROUTERS` whenever `settings.py:REACTPY_DATABASE` is set.
- Added system checks for a variety of common ReactPy misconfiguration

+++ src/reactpy_django/utils.py
@@ -332,12 +332,7 @@ def create_cache_key(*args):
 def db_cleanup(immediate: bool = False):
     """Deletes expired component sessions from the database.
     This function may be expanded in the future to include additional cleanup tasks."""
-    from .config import (
-        REACTPY_CACHE,
-        REACTPY_DATABASE,
-        REACTPY_DEBUG_MODE,
-        REACTPY_RECONNECT_MAX,
-    )
+    from .config import REACTPY_CACHE, REACTPY_DEBUG_MODE, REACTPY_RECONNECT_MAX
     from .models import ComponentSession
 
     clean_started_at = datetime.now()
@@ -351,7 +346,7 @@ def db_cleanup(immediate: bool = False):
     expires_by: datetime = timezone.now() - timedelta(seconds=REACTPY_RECONNECT_MAX)
 
     # Component params exist in the DB, but we don't know when they were last cleaned
-    if not cleaned_at_str and ComponentSession.objects.using(REACTPY_DATABASE).all():
+    if not cleaned_at_str and ComponentSession.objects.all():
         _logger.warning(
             "ReactPy has detected component sessions in the database, "
             "but no timestamp was found in cache. This may indicate that "
@@ -361,9 +356,7 @@ def db_cleanup(immediate: bool = False):
     # Delete expired component parameters
     # Use timestamps in cache (`cleaned_at_str`) as a no-dependency rate limiter
     if immediate or not cleaned_at_str or timezone.now() >= clean_needed_by:
-        ComponentSession.objects.using(REACTPY_DATABASE).filter(
-            last_accessed__lte=expires_by
-        ).delete()
+        ComponentSession.objects.filter(last_accessed__lte=expires_by).delete()
         caches[REACTPY_CACHE].set(cache_key, now_str, timeout=None)
 
     # Check if cleaning took abnormally long
